+++ server/standalone/dbal/computeplan.go
@@ -35,14 +35,14 @@ func (d *DBAL) GetComputePlan(key string) (*asset.ComputePlan, error) {
 	builder := pgDialect.Select().
 		Column("cp.asset").
 		Column(squirrel.Expr("COUNT(t.id)")).
-		Column(squirrel.Expr("COUNT(t.id) FILTER (WHERE t.asset->>'status' = 'STATUS_WAITING')")).
-		Column(squirrel.Expr("COUNT(t.id) FILTER (WHERE t.asset->>'status' = 'STATUS_TODO')")).
-		Column(squirrel.Expr("COUNT(t.id) FILTER (WHERE t.asset->>'status' = 'STATUS_DOING')")).
-		Column(squirrel.Expr("COUNT(t.id) FILTER (WHERE t.asset->>'status' = 'STATUS_CANCELED')")).
-		Column(squirrel.Expr("COUNT(t.id) FILTER (WHERE t.asset->>'status' = 'STATUS_FAILED')")).
-		Column(squirrel.Expr("COUNT(t.id) FILTER (WHERE t.asset->>'status' = 'STATUS_DONE')")).
+		Column(squirrel.Expr("COUNT(t.id) FILTER (WHERE t.status = 'STATUS_WAITING')")).
+		Column(squirrel.Expr("COUNT(t.id) FILTER (WHERE t.status = 'STATUS_TODO')")).
+		Column(squirrel.Expr("COUNT(t.id) FILTER (WHERE t.status = 'STATUS_DOING')")).
+		Column(squirrel.Expr("COUNT(t.id) FILTER (WHERE t.status = 'STATUS_CANCELED')")).
+		Column(squirrel.Expr("COUNT(t.id) FILTER (WHERE t.status = 'STATUS_FAILED')")).
+		Column(squirrel.Expr("COUNT(t.id) FILTER (WHERE t.status = 'STATUS_DONE')")).
 		From("compute_plans AS cp").
-		LeftJoin("compute_tasks AS t ON (t.asset->>'computePlanKey')::uuid = cp.id").
+		LeftJoin("compute_tasks AS t ON t.compute_plan_key = cp.id").
 		Where(squirrel.Eq{"cp.id": key}).
 		Where(squirrel.Eq{"cp.channel": d.channel}).
 		GroupBy("cp.id")
@@ -106,14 +106,14 @@ func (d *DBAL) QueryComputePlans(p *common.Pagination, filter *asset.PlanQueryFi
 	builder := pgDialect.Select().
 		Column("cp.asset").
 		Column(squirrel.Expr("COUNT(t.id)")).
-		Column(squirrel.Expr("COUNT(t.id) FILTER (WHERE t.asset->>'status' = 'STATUS_WAITING')")).
-		Column(squirrel.Expr("COUNT(t.id) FILTER (WHERE t.asset->>'status' = 'STATUS_TODO')")).
-		Column(squirrel.Expr("COUNT(t.id) FILTER (WHERE t.asset->>'status' = 'STATUS_DOING')")).
-		Column(squirrel.Expr("COUNT(t.id) FILTER (WHERE t.asset->>'status' = 'STATUS_CANCELED')")).
-		Column(squirrel.Expr("COUNT(t.id) FILTER (WHERE t.asset->>'status' = 'STATUS_FAILED')")).
-		Column(squirrel.Expr("COUNT(t.id) FILTER (WHERE t.asset->>'status' = 'STATUS_DONE')")).
+		Column(squirrel.Expr("COUNT(t.id) FILTER (WHERE t.status = 'STATUS_WAITING')")).
+		Column(squirrel.Expr("COUNT(t.id) FILTER (WHERE t.status = 'STATUS_TODO')")).
+		Column(squirrel.Expr("COUNT(t.id) FILTER (WHERE t.status = 'STATUS_DOING')")).
+		Column(squirrel.Expr("COUNT(t.id) FILTER (WHERE t.status = 'STATUS_CANCELED')")).
+		Column(squirrel.Expr("COUNT(t.id) FILTER (WHERE t.status = 'STATUS_FAILED')")).
+		Column(squirrel.Expr("COUNT(t.id) FILTER (WHERE t.status = 'STATUS_DONE')")).
 		From("compute_plans AS cp").
-		LeftJoin("compute_tasks AS t ON (t.asset->>'computePlanKey')::uuid = cp.id").
+		LeftJoin("compute_tasks AS t ON t.compute_plan_key = cp.id").
 		Where(squirrel.Eq{"cp.channel": d.channel}).
 		GroupBy("cp.id").
 		OrderBy("cp.asset->>'creationDate' ASC", "cp.id ASC").