+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - return `datasamples` list in `RegisterDataSamplesResponse` (#486)
 - return `tasks` list in `RegisterTasksResponse` (#493)
 - store the error type of a failed compute task in a failure report instead of an event (#487)
+- improve performance of `compute_tasks` SQL indexes by using dedicated columns instead of JSONB (#503)
 
 ### Fixed
 - set the correct name of the `RegisterFailureReport` service method used in distributed mode (#485)